Added config for redirecting to patient chart Dashboard
