--- packages/esm-patient-search-app/src/compact-patient-search-extension/patient-search-bar.scss
@@ -1,19 +1,0 @@
-@use '@carbon/styles/scss/colors';
-@use '@carbon/styles/scss/spacing';
-@import '~@openmrs/esm-styleguide/src/vars';
-
-.searchArea {
-  width: inherit;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border: 1px solid $ui-04;
-}
-
-.patientSearchInput {
-  border: none;
-}
-
-.patientSearchInput input:focus {
-  outline: 1px solid 1px solid colors.$orange-40;
-}
