+++ packages/esm-patient-search-app/src/config-schema.ts
@@ -6,6 +6,12 @@ export const configSchema = {
       _description: 'Where clicking a patient result takes the user. Accepts template parameter ${patientUuid}',
       _validators: [validators.isUrlWithTemplateParameters(['patientUuid'])],
     },
+    redirectToPatientDashboard: {
+      _type: Type.String,
+      _default: 'Patient Summary',
+      _description:
+        'On clicking the patient banner in the search results, which should be the default patient chart dashboard to redirect to.',
+    },
   },
   includeDead: {
     _type: Type.Boolean,
