+++ packages/esm-patient-search-app/src/patient-search-page/patient-search-lg.component.tsx
@@ -56,7 +56,7 @@ const PatientSearchComponent: React.FC<PatientSearchComponentProps> = ({
         navigate({
           to: `${interpolateString(config.search.patientResultUrl, {
             patientUuid: patientUuid,
-          })}/${encodeURIComponent('Patient Summary')}`,
+          })}/${encodeURIComponent(config.search.redirectToPatientDashboard)}`,
         });
       }
       if (hidePanel) {
